% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{conTocat}
\alias{conTocat}
\title{Convert continuous values to categorical/discrete values}
\usage{
conTocat(covariates, var_names)
}
\arguments{
\item{covariates}{samples x covariates matrix containing the continous covariates which need to be discretized}

\item{var_names}{a list of the continuous variables that need to be discretized}
}
\description{
This function calculates the quantiles for the list of input values and assigns a quantile group to every data point
}
\note{
the function returns a modified  samples x covariates matrix with the categorized variables appended to 
the end as another column
}
\author{
Karthikeyan Murugesan<karthikeyanm60 at yahoo.com>
#@examples conTocat(covariates, var_names)
}
\keyword{categorize,}
\keyword{discretize}
\keyword{quantile,}

