\name{ExpressionNormalizationWorkflow-package}
\alias{ExpressionNormalizationWorkflow}
\docType{package}
\title{
Gene Expression Normalization Workflow
}
\description{
An extensive, customized expression normalization workflow
incorporating Supervised Normalization of Microarryas(SNM), Surrogate Variable
Analysis(SVA) and Principal Variance Component Analysis to identify batch
effects and remove them from the expression data to enhance the ability to
detect the underlying biological signals.
}
\details{
\tabular{ll}{
Package: \tab ExpressionNormalizationWorkflow\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-30\cr
License: \tab GPL (>=3)\cr
}
}
\author{
Karthikeyan Murugesan<karthikeyanm60 at yahoo.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
#~~ simple examples of the most important functions ~~
}
