% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdaobj.R
\name{expSetobj}
\alias{expSetobj}
\title{Create a compact ExpressionSet Datatype Object}
\usage{
expSetobj(expression, covariates, cov_desc = NULL)
}
\arguments{
\item{expression}{the gene expression value martix with features (probes)  as rows and samples as columns}

\item{covariates}{the phenotype (covariate) matrix with samples as rows and the covariates as columns}

\item{cov_desc}{a list of descriptors for each of the covariates}
}
\value{
exp_datObj ExpressionSet Datatype Object
}
\description{
An expression set object can store the expression values and the 
meta data associated like the covariate information, annotations and experiment protocol data
}
\note{
The return values of the function can be accessed as variable$parameter
}
\author{
Karthikeyan  Murugesan <karthikeyanm60 at yahoo.com>
}
\references{
Gentleman RC, Carey VJ, Bates DM and others (2004).
“Bioconductor: Open software development for computational biology and bioinformatics.” Genome Biology, 5, pp. R80.
#@examples expSetobj(expression, covariates)
}
\keyword{ExpressionSet}

